backend/requirements.txt
fastapi[standard]~=0.115.5
sqlmodel~=0.0.22
pyjwt~=2.10.0
passlib[bcrypt]~=1.7.4
python-dotenv~=1.0.1
psycopg2-binary~=2.9.10

pydantic~=2.9.2
uvicorn~=0.32.0
SQLAlchemy~=2.0.36
alembic~=1.14.0
pytest~=8.3.3
httpx~=0.28.1
loguru